+++ src/minify/minify.js
@@ -2,8 +2,20 @@
 
 const imagemin = require("imagemin");
 
-function minify(input, options) {
-  const { imageminOptions, bail } = options;
+function minify(task = {}) {
+  const { imageminOptions, bail, input } = task;
+  const result = {
+    errors: new Set(),
+    output: null,
+    warnings: new Set()
+  };
+
+  if (!input) {
+    result.errors.add(new Error("Empty input"));
+
+    return result;
+  }
+
   // Ensure that the contents i have are in the form of a buffer
   const source = Buffer.isBuffer(input) ? input : Buffer.from(input);
   const originalSize = source.length;
@@ -13,7 +25,9 @@ function minify(input, options) {
     !imageminOptions.plugins ||
     imageminOptions.plugins.length === 0
   ) {
-    throw new Error("No plugins found for `imagemin`");
+    result.errors.add(new Error("No plugins found for `imagemin`"));
+
+    return result;
   }
 
   // Await for imagemin to do the compression
@@ -22,17 +36,21 @@ function minify(input, options) {
       .buffer(source, imageminOptions)
       .then(optimizedSource => {
         if (optimizedSource.length < originalSize) {
-          return optimizedSource;
+          result.output = optimizedSource;
+        } else {
+          result.output = source;
         }
 
-        return input;
+        return result;
       })
       .catch(error => {
         if (bail) {
-          throw error;
+          result.errors.add(error);
+        } else {
+          result.warnings.add(error);
         }
 
-        return Promise.resolve(source);
+        return result;
       })
   );
 }