+++ CHANGELOG.md
@@ -6,8 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org).
 
 ## HEAD
 
+* Changed: use `webpack` errors and warnings `api`.
 * Changed: `imagemin` plugins for `loader` should be placed in `imageminOptions` options (look example in `README.md`).
-* Chore: refactor all plugin (all source code now in `src` directory).
+* Chore: refactor all plugin code (all source code now in `src` directory).
 * Fixed: use `maxConcurrency`, use `os.cpus().length - 1` for multi core systems and `1` for single core system.
 * Fixed: don't add same file paths in `manifest`.